target-arm: Support v6 barriers in linux-user mode
ARMv6 implemented various operations as special cases of cp15 accesses
which are true instructions in v7; this includes barriers (DMB, DSB, ISB).
Catch this special case at translate time, so that it works in linux-user
mode (which doesn't provide a functional get_cp15 helper) as well as
system mode.

Includes minor cleanup of the existing cases, and moving the check to
before the "OK in user mode?" test, since we know all these instructions
are OK.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>