target-arm: Convert cp15 crn=9 registers
Convert cp15 crn=9 registers (mostly cache lockdown)
to the new scheme.

Note that this change makes OMAPCP cores RAZ/WI the whole
c9 space. This is a change from previous behaviour, but a
return to the behaviour of commit c3d2689d when OMAP1
support was first added -- subsequent commits have clearly
accidentally relegated the OMAPCP RAZ condition to only
a subset of the crn=9 space when adding support for other
cores.

TODO: doesn't this disable the c15 dummy ?

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>