hw/omap_intc.c: Don't complain about nonexistent register read by Linux
The Linux omap kernel now deliberately reads from the nonexistent register
at 0xf8, which only exists on ti816x, because it knows that on omap2 and
omap3 this will read as zero. Handle it that way (ie suppress qemu's
complaint about read of a bad register), since otherwise the omap3
model has so many warnings it is unusable.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>