target-arm: Fault on invalid first level descriptor type
The low two bits of a first level page table entry specify
its type. In v7 VMSA where the PXN attribute is unsupported
(which is true for QEMU at the moment) the type 0b11 is
reserved and generates a translation fault in the same way
as the 0b00 invalid type.

(In VMSAv6 type 0b11 is RESERVED; prior to that it was
used for fine second level page tables. Although it isn't
stated in the ARM ARM the way it is for v7, the 1136,
1176 and 11MPCore TRMs all state that 0b11 causes a
translation fault in the v6 page table format.)

TODO: this patch is OK, but maybe we should just implement PXN?

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>