linux-headers: Add ARM KVM headers (not for upstream)
This commit adds the ARM KVM headers. This is not to go to QEMU
upstream -- the correct path there is that the KVM code will be
committed to a mainline upstream kernel, and then upstream QEMU
can do a bulk header update from the upstream kernel, which will
allow us to drop this temporary commit.

This includes a dummy kvm_sync_regs {} since QEMU has picked
up a newer version of the general KVM headers which are a bit
ahead of the ARM KVM headers for the moment, so it isn't a
pure copy of any ARM KVM kernel header set. This wrinkle will
go away at some point when ARM KVM rebases on kvm.next and
we do a cleaner import, but for now we need this just to avoid
a compilation failure.

This commit is based on the headers as of commit 827ee5b in
git://github.com/virtualopensystems/linux-kvm-arm kvm-a15-v7