ARM: KVM: Add support for KVM on ARM architectures
This patch represents work in progress on adding support
for KVM on ARM architectures. Note that it is highly
experimental and is not enabled by default.

TODO: this is Christoffer's patch, but with the configure
script changes redone by me. Needs signed-off-by.
Has also had a few subsequent fixes squashed in.
Consider whether this can split apart into separate patches?