kvm: Decouple 'interrupt injection is async' from 'kernel irqchip'
On x86 interrupt injection is asynchronous (and therefore
VCPU idle management is done in the kernel) if and only
if there is an in-kernel irqchip. On other architectures this
isn't necessarily true (they may always do asynchronous
injection), so define a new kvm_async_interrupt_injection()
function instead of misusing kvm_irqchip_in_kernel().

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>