target-arm: Make our encoding of cp regs match the kernel's
Both QEMU and the kernel have an arbitrary mapping of the
(cp,is64,crn,crm,op1,op2) tuple to an integer index. Unfortunately
they picked slightly different layouts. Change QEMU's mapping
(knowledge of which is confined to the ENCODE_CP_REG() macro)
to match the kernel's; this avoids having to do a conversion
in code to save and restore registers from the kernel.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>