versatile_pci: Drop 'realview_pci'
There isn't really any difference between the PCI controller on
the VersatilePB and the one on the RealviewEB -- they're both
the same Xilinx PCI controller. (The Realview PB-A8 and PBX-A9
do have a different PCI controller, but we don't model that at all.)
Comments in QEMU claim that the PB version doesn't support PCI
I/O, but this is incorrect. Drop the unnecessarily different
'realview_pci' class.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>