hw/versatile_pci: Implement the correct PCI IRQ swizzling
Implement the correct IRQ swizzling for the Versatile PCI controller
(it is the standard PCI one). Note that this change will break
Linux kernels which rely on the broken QEMU behaviour!

XXX not quite true -- it will break if they have more than
one PCI device -- slot B's irq mapping stays the same but
others break. [similarly true for all slots which == 2 mod 4]

TODO: making the controller slot 29 and the min slot 29
is correct for h/w but limits qemu to two PCI devices,
which is bad news. Consider changing back to 11...

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>