target-arm: Add migration-only interface to cache ID registers
The cache ID registers have an interface where one register
is used to select which of the array of underlying cache ID
registers to read, and then another register is used to
actually read the value. For migration we need direct access
to the underlying registers; provide ARMCPRegInfo definitions
which give this direct access. These line up with the KVM
kernel interface for cache ID register access; since the
register ID is not encodable in the MSR/MRS instruction
these migration-only definitions don't appear to the guest.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>