AARCH64: Add boot support for aarch64 processor
This version supports booting of a single Aarch64 CPU by setting appropriate
registers. The bootloader includes placehoders for Board-ID that are used to
implementing uniform indexing across different bootloaders. The same macro
names are used with different values when compiling for different processors.

Signed-off-by: Mian M. Hamayun <m.hamayun@virtualopensystems.com>

Conflicts:

	hw/arm/boot.c

Conflicts:

	hw/arm/boot.c

[PMM notes: this patch is pretty minimal, which is nice, but
I want to think about whether there's a nicer way to do any
of this.]