AARCH64: Add aarch64 CPU initialization, get and put registers support
The cpu init function tries to initialize with all possible cpu types, as
KVM does not provide a means to detect the real cpu type and simply refuses
to initialize on cpu type mis-match. By using the loop based init function,
we avoid the need to modify code if the underlying platform is different,
such as Fast Models instead of Foundation Models.

Get and Put Registers deal with the basic state of Aarch64 CPUs for the moment.

Signed-off-by: Mian M. Hamayun <m.hamayun@virtualopensystems.com>

Conflicts:

	target-arm/kvm.c

Conflicts:

	target-arm/cpu.c

Conflicts:

	linux-headers/linux/kvm.h

[PMM: squashed the #ifdef around aarch32 specific reset code into this patch
Other notes:
 * the next patch in this branch fixes up stuff in this patch; should
   be folded/rearranged instead
 * we should have a kvm32.c/kvm64.c rather than these ifdefs
 * we'll need to deal with system registers at some point, including
   proper cpu reset
]