BUGFIX disas_add_sub_imm by rewriting it
disas_add_sub_imm was broken in several ways (starting with
erroneously negating the immediate rather than inverting it
for subtract operations, but there was at least one other bug
too.)

In particular this is why the libc "check the kernel version"
code was erroneously barfing.

Rather than try to track down what it was doing wrong, just
rewrite it so it does the obvious sensible thing of "just emit
an add/sub if this isn't a flagsetting operation" and has
a reasonably efficient specialcasing of the flagsetting ops.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>