BUGFIX offset calculation in ldp
The imm7 offset field in LDP needs to be scaled according to the size:
this was being done in STP but not LDP... (The most visible effect of
this was that after a postfix operation like "LDP x29,x30,[sp],#496"
the stack pointer would end up misaligned.)

ALso put the offset in an int64_t, not just an int. In fact
the immediate size and scale mean we can't lose bits here if
int is 32 bit, but it's clearer to use int64_t since the
pseudocode specifies 64 bit arithmetic for addresses.

-- PMM