softfloat: Fix float64_to_uint64
The comment preceding the float64_to_uint64 routine suggests that
the implementation is broken.  And this is, indeed, the case.

This patch properly implements the conversion of a 64-bit floating
point number to an unsigned, 64 bit integer.

This contribution can be licensed under either the softfloat-2a or -2b
license.

V2: Added softfloat license statement.

V3: Modified to meet QEMU coding conventions.

V4: Fixed incorrect handling of small negatives, which, if rounded
up to zero should not set the inexact flag.

Signed-off-by: Tom Musta <tommusta@gmail.com>
Message-id: 1387397961-4894-2-git-send-email-tommusta@gmail.com

PMM notes: this looks like it gives the wrong answer for NaNs
with the sign bit set.