hw/arm/ranchu: New machine model for 64-bit ARM Android emulator
Add a new machine model for the 64-bit ARM Android emulator.
This is based heavily on the 'virt' machine model, and like
that we also create a device tree to pass to the guest kernel.
The major difference is that this board will have all the
Android-specific devices in it.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>