android_pipe: import the qemu_pipe/goldfish_pipe code
This brings in the android emulator pipe code from the current aosp-qemu
master branch. So far it's just a comment and hack to get compiling
change:

  - remove dependency on android utils (ANEW/ASTRDUP/AFREE)
  - detach from the goldfish_device bus bits (inc irq raising)
  - import a few helper functions (uint64_set_high/low, goldfish_guest_is_64bit)
  - disable the VM save/restore code
  - use current_cpu for memory translation ops instead of cpu_single_env
  - fix debugging prints using portable formats
  - common code path for translating vaddr->qemu addr*
  - make some debug statments user-visible (unimp and guest errors)

There is a hacking backlink between AndroidPipeState and the PipeDevice
stuff (which should be merged cleanly later) just so we can get back to
->irq when we need to.

* I commonised the code and used contemporary qemu functions. However for
upstream the kernel driver need to move to passing hwaddr's to
QEMU (pref using virt-io mechanisms).