android_adb: Provide adb_server_init() function
This should be squashed into an earlier patch.

Instead of initializing the backend unconditionally from the
android_pipe init with a hardcoded port, just provide a
function to be called by the board which does the backend
init with a passed in port number, and returns false if it
could not listen on that port.