linux-user: Range check the nfds argument to ppoll syscall
Do an initial range check on the ppoll syscall's nfds argument,
to avoid possible overflow in the calculation of the lock_user()
size argument. The host kernel will later apply the rather lower
limit based on RLIMIT_NOFILE as appropriate.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
---
This fixes an LTP test case which passes -1 for nfds.
The test is kind of bogus in that it should ideally
pass a valid-but-over-the-rlimit nfds for this corner
case rather than assuming EINVAL is checked before
EFAULT, but in any case QEMU should protect itself
against the overflow.