tests/virtio-9p-test: Don't call le*_to_cpus on fields of packed struct
For a packed struct like 'P9Hdr' the fields within it may not be
aligned as much as the natural alignment for their types.  This means
it is not valid to pass the address of such a field to a function
like le32_to_cpus() which operate on uint32_t* and assume alignment.
Doing this results in a SIGBUS on hosts like SPARC which have strict
alignment requirements.

Use ldl_le_p() instead, which is specified to correctly handle
unaligned pointers.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
---
Sadly gcc doesn't warn about this:
 https://gcc.gnu.org/bugzilla/show_bug.cgi?id=51628
clang supposedly was fixed to do so:
 https://bugs.llvm.org//show_bug.cgi?id=22821
but I think that commit was reverted without the bug being
reopened; at least my clang doesn't have that warning flag.