arm: Implement HFNMIENA support for M profile MPU
Implement HFNMIENA support for the M profile MPU. This bit controls
whether the MPU is treated as enabled when executing at execution
priorities of less than zero (in NMI, HardFault or with the FAULTMASK
bit set).

Doing this requires us to use a different MMU index for "running
at execution priority < 0", because we will have different
access permissions for that case versus the normal case.


TODO:
 figure out mmu index usage for M profile -- look at
 v8M while we're here
 identify what bits of code care about mmu index
 set mmu index differently for negative-exec-prio
 make regime_translation_disabled honour HFMNMIENA