configure: Consistently define _XOPEN_SOURCE=700
The ncurses header file requires _XOPEN_SOURCE to be defined to
cause it to provide some wide character functions we use.
With glibc this happens automatically, because if _GNU_SOURCE
is defined then glibc defines _XOPEN_SOURCE. For other libc
implementations (like musl and the OSX libc) this does not
happen, and so we need to define _XOPEN_SOURCE ourselves to
request these functions.

Awkwardly, defining _XOPEN_SOURCE also means "and don't
provide any functions that aren't in that standard" unless
we also define some libc-specific define to enable those
extra functions again:
 * glibc and musl: _GNU_SOURCE (already set)
 * Solaris: __EXTENSIONS__ (already set)
 * OSX: _DARWIN_C_SOURCE
 * BSDs: ???

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>